# openapi.yaml
openapi: 3.1.0
info:
  title: Self Agent ID API
  version: 1.0.0
  description: >
    On-chain AI agent identity registry with proof-of-human verification.
    Register, deregister, and query agents — no API keys required.
    Sessions use encrypted tokens with a 30-minute TTL.
  contact:
    name: Self Protocol
    url: https://github.com/selfxyz/self-agent-id
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://self-agent-id.vercel.app
    description: Production
...
security: []
...
components:
  securitySchemes:
    selfAgentAuth:
      type: apiKey
      in: header
      name: x-self-agent-signature
      description: >
        ECDSA signature produced by `SelfAgent.signRequest()`.
        Must be accompanied by `x-self-agent-address` and
        `x-self-agent-timestamp` headers.
    aaProxyToken:
      type: apiKey
      in: header
      name: x-aa-proxy-token
      description: >
        Short-lived token issued by POST /api/aa/token.
        Required for bundler and paymaster proxy endpoints.
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Credentials:
      type: object
      properties:
        nationality:
          type: string
          description: ISO 3166-1 alpha-3 country code
        olderThan:
          type: integer
        ofac:
          type: array
          items:
            type: boolean
          description: "[isOnSdnList, isOnConsolidatedList, isOnOfacList]"
...
    DemoVerifyResponse:
      type: object
      properties:
        valid:
          type: boolean
        agentAddress:
          type: string
        agentKey:
          type: string
        agentId:
          type: string
        agentCount:
          type: string
        verificationCount:
          type: integer
        credentials:
          $ref: "#/components/schemas/Credentials"
        error:
          type: string
...
    ChainVerifyRequest:
      type: object
      required: [agentKey, nonce, deadline, eip712Signature]
      properties:
        agentKey:
          type: string
          description: Agent's bytes32 public key
        nonce:
          type: string
        deadline:
          type: integer
          description: Unix timestamp
        eip712Signature:
          type: string
        networkId:
          type: string
          enum: [celo, celo-sepolia]
    ChainVerifyResponse:
      type: object
      properties:
        txHash:
          type: string
        blockNumber:
          type: integer
        explorerUrl:
          type: string
        agentAddress:
          type: string
        agentId:
          type: string
        credentials:
          type: object
          properties:
            olderThan:
              type: string
            nationality:
              type: string
        verificationCount:
          type: string
        totalVerifications:
          type: string
        gasUsed:
          type: string
        rateLimitRemaining:
          type: integer
        pending:
          type: boolean
          description: "True if tx submitted but confirmation timed out"
...
paths:
...
  /api/demo/verify:
    post:
      tags: [Demo]
      summary: Verify agent identity (service pattern)
      operationId: demoVerify
      description: >
        Accepts a signed agent request, verifies the ECDSA signature and
        on-chain status, then returns agent info and credentials.
      security:
        - selfAgentAuth: []
      parameters:
        - name: network
          in: query
          schema:
            type: string
            enum: [celo, celo-sepolia]
          description: Network override (defaults to celo-sepolia)
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DemoVerifyResponse"
        "401":
          description: Missing agent authentication headers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Replay detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/demo/chain-verify:
    post:
      tags: [Demo]
      summary: On-chain meta-transaction verification
      operationId: demoChainVerify
      description: >
        Relayer endpoint that submits EIP-712 meta-transactions to the
        `AgentDemoVerifier` contract. Rate limited to 3 per hour per
        human nullifier.
      security:
        - selfAgentAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChainVerifyRequest"
      responses:
        "200":
          description: Transaction submitted and confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChainVerifyResponse"
        "400":
          description: Simulation failed or invalid request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing agent authentication headers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Agent verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Replay detected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limited (3 per hour per human)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  rateLimitRemaining:
                    type: integer
                  retryAfterMs:
                    type: integer
...